third_party/blink/web_tests/fast/encoding/invalid-xml-expected.txt

Testing: invalid-xml-utf8.xml
PASS iframe.contentDocument.documentElement.tagName is "root"
PASS iframe.contentDocument.documentElement.getElementsByTagName('test').length < 1 is true
Testing: invalid-xml-utf16.xml
PASS iframe.contentDocument.documentElement.tagName is "root"
PASS iframe.contentDocument.documentElement.getElementsByTagName('test').length < 1 is true
Testing: invalid-xml-shift-jis.xml
PASS iframe.contentDocument.documentElement.tagName is "root"
PASS iframe.contentDocument.documentElement.getElementsByTagName('test').length < 1 is true
Testing: invalid-xml-x-mac-thai.xml
PASS iframe.contentDocument.documentElement.tagName is "root"
PASS iframe.contentDocument.documentElement.getElementsByTagName('test').length < 1 is true
PASS successfullyParsed is true

TEST COMPLETE
